apps/llm-worker/pyproject.toml
[project]
name = "llm-worker"
version = "1.0.0"
description = "Default template for PDM package"
authors = [
    {name = "Gustavo Henrique", email = "ghfc2000hotmail.com"},
]
dependencies = ["transformers>=4.53.0", "accelerate>=1.8.1", "python-dotenv>=1.1.1", "hf-xet>=1.1.5", "pandas>=2.3.1", "editdistance>=0.8.1", "torch>=2.7.1", "torchvision>=0.22.1", "torchaudio>=2.7.1"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
distribution = false

[tool.pdm.scripts]
initialize = "python src/initialize.py"
start = "python src/main.py"

[tool.pdm.resolution]
respect-source-order = true
 
[[tool.pdm.source]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
include_packages = ["torch", "torchvision", "torchaudio"]
verify_ssl = true

[[tool.pdm.source]]
name = "pypi"
url = "https://pypi.org/simple"